+++ GeneralsMD/Code/GameEngineDevice/Include/VideoDevice/Bink/BinkVideoPlayer.h
@@ -131,7 +131,12 @@ class BinkVideoPlayer : public VideoPlayer
 		virtual VideoStreamInterface*	load( AsciiString movieTitle );	///< Load video file in to memory for playback
 
 		virtual void notifyVideoPlayerOfNewProvider( Bool nowHasValid );
+		virtual void setVolume( Real volume );
 		virtual void initializeBinkWithMiles( void );
+
+	private:
+
+		static Int movieAudioVolume( Real volume );
 };